refactor: GetReservationsByUserIdAsync to GetReservationsWithAccommodationByUserIdAsync

+++ Restino.Application/Features/Reservations/Commands/DeleteReservation/DeleteReservationCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using MediatR;
 using Restino.Application.Contracts.Persistance;
+using Restino.Application.Exceptions;
 using Restino.Domain.Entities;
 
 namespace Restino.Application.Features.Reservations.Commands.DeleteReservation
@@ -17,6 +18,10 @@ namespace Restino.Application.Features.Reservations.Commands.DeleteReservation
         public async Task<Unit> Handle(DeleteReservationCommand request, CancellationToken cancellationToken)
         {
             var reservationToDelete = await _reservationRepository.GetByIdAsync(request.Id);
+
+            if (reservationToDelete == null)
+                throw new NotFoundException(nameof(Accommodation), request.Id);
+
             await _reservationRepository.DeleteAsync(reservationToDelete);
 
             return Unit.Value;
